Mutations now need authorization in the headers

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,53 +1,69 @@
 import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
-        const emailTaken = await prisma.exists.User({ email: args.data.email })
-
-        if (emailTaken) {
-            throw new Error('Email taken')
-        }
-
-        if(args.data.password.length < 3) {
-            throw new Error('Must be 3 character or longer')
+        if (args.data.password.length < 8) {
+            throw new Error('Password must be 8 characters or longer.')
         }
 
         const password = await bcrypt.hash(args.data.password, 10)
-
-        return prisma.mutation.createUser({
+        const user = await prisma.mutation.createUser({
             data: {
                 ...args.data,
                 password
             }
-        }, info)
-    },
-    deleteUser(parent, args, { prisma }, info) {
-        const userExists = prisma.exists.User({ id: args.id })
-        if (!userExists) {
-            throw new Error('User not found')
+        })
+
+        return {
+            user,
+            token: jwt.sign({ userId: user.id }, 'thisisasecret')
         }
+    },
+    async login(parent, args, { prisma }, info) {
+        const user = await prisma.query.user({
+            where: {
+                email: args.data.email
+            }
+        })
+
+        if (!user) {
+            throw new Error('Unable to login')
+        }
+
+        const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+        if (!isMatch) {
+            throw new Error('Unable to login')
+        }
+
+        return {
+            user,
+            token: jwt.sign({ userId: user.id }, 'thisisasecret')
+        }
+    },
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
 
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    updateUser(parent, args, { prisma }, info) {
-        // We can not check if this user exists but we get a bad error message
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
-            data: args.data 
-        }, info) 
+            data: args.data
+        }, info)
     },
-    createPost(parent, args, { prisma, pubsub }, info) {
-        const userExists = prisma.exists.User({ id: args.data.id })
-
-        if (!userExists) {
-            throw new Error('User not found')
-        }
+    createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
 
         return prisma.mutation.createPost({
             data: {
@@ -56,46 +72,60 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         }, info)
     },
-    deletePost(parent, args, { prisma, pubsub }, info) {
-        const postExists = prisma.exists.Post({ id: args.id})
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
 
         if (!postExists) {
-            throw new Error('Post not found')
+            throw new Error('Unable to delete post')
         }
-
+        
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updatePost(parent, args, { prisma, pubsub }, info) {
-        const postExists = prisma.exists.Post({ id: args.id })
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
 
         if (!postExists) {
-            throw new Error('Post not found')
+            throw new Error('Unable to update post')
         }
 
         return prisma.mutation.updatePost({
-            data: args.data,
             where: {
                 id: args.id
-            }
+            },
+            data: args.data
         }, info)
     },
-    createComment(parent, args, { prisma }, info) {
+    createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -106,14 +136,38 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
